+++ ui/src/components/layer/upload/layer-upload-kml-modal.ts
@@ -72,7 +72,7 @@ export class LayerUploadKmlModal extends LitElementI18n {
     }
 
     h2 {
-      ${applyTypography('modal-title')}
+      ${applyTypography('modal-title-1')}
       color: var(--color-text--emphasis--high);
       margin: 0;
     }
