+++ ui/src/components/layout/index.ts
@@ -1,0 +1,2 @@
+import './language-selector';
+import './tracking-consent-modal';
