+++ ui/src/styles/theme.ts
@@ -9,7 +9,8 @@ type Typography =
   | 'title'
   | 'subtitle-1'
   | 'subtitle-2'
-  | 'modal-title'
+  | 'modal-title-1'
+  | 'modal-title-2'
   | 'button'
   | 'caption'
   | 'overline'
